requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
python_docx
Pillow